calibration for the preventive model
LHC sampling model selection for the preventive model. We still need to
decide how to define a preventive model. Firstly, if the entropy is
doing anything interesting (halo gas distribution) for the metal
distribution. Second, whether certain amount of outflow is still
allowed or treated a freedom in the calibration.